commerceapp/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.44.0
commerce==0.1.dev0
Django==5.2
django-ckeditor==6.7.2
django-js-asset==3.1.2
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
